08-HandOff_Agent_As_ToolCall/pyproject.toml
[project]
name = "handoff-agent-as-toolcall-panacloude"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai-agents>=0.0.14",
    "python-dotenv>=1.1.0",
]
